+++ openeo/udf/__init__.py
@@ -5,6 +5,7 @@ class OpenEoUdfException(BaseOpenEoException):
     pass
 
 
+from openeo.udf.debug import inspect
 from openeo.udf.feature_collection import FeatureCollection
 from openeo.udf.run_code import run_udf_code, execute_local_udf
 from openeo.udf.structured_data import StructuredData